Updating how images get loaded.
Changes for "im_feeling_lucky" page.
Feature 1:
- Instead of having a default image shown, the code now waits for the
page to be loaded. Then if the user has not provided any URL hash
prompt, then a default image is loaded.

- This eliminates the problem of showing an image then, updating that
image to the provided URL hash, which looks kind of glitchy.

Feature 2:
- Removed AWS dependency from bundle.js
- Previous the AWS SDK was used for getting the complete list of images
stored in the s3 bucket.
- This worked okay but the AWS SDK was 6mb when bundled into the page
which is HUGE!
- Instead create a JSON file that has all the image prompt paths stored
in it and the image paths get loaded from that file.
- The `2022_prompt_paths.json` file is 10kb
- The new bundle.js file is 5kb down from 6.7mb (some 1000%(s) times
smaller, didn't compute it lol)

Updated README with details about how to run the script to generate `2022_prompt_paths.json`